Add a fuzzing mode that makes the decoding NOP To try and fuzz original raw files fast add a new decoding mode where the actual image decoding is skipped. This should exercise pretty much all the metadata handling across the code but just not run the actual conversion. In the future this can also be useful as a metadata only mode to rawloader so you can grab all the metadata from the code quickly without having to wait potentially up to 700ms for image decoding you don't need. The only downside would be no longer having calculated black levels in cameras with masked areas in the sensor. Possibly for those cameras the NOP mode would need to be ignored just to have less surprises.
Preliminary testing shows that the formats should now be more than fast
enough for fuzzing as they run on the order of 10-100k iterations per
second if the file is already in memory. Unfortunately the normal afl
mode is to feed in a new file over stdin so that process is currently
making any reasonable fuzzing speed impossible. The very quick test runs
I did showed worst cases of 3/sec and even the more reasonable cases
were still 100/sec.